# src/main/java/vasyurin/work/bank_rest/dto/UserDto.java
package vasyurin.work.bank_rest.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.sql.Timestamp;


@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class UserDto {

    private Long id;
    private String username;
    private String role;
    private Timestamp created;
    private Timestamp updated;

}