# src/main/java/vasyurin/work/bank_rest/dto/UserRequestDto.java
package vasyurin.work.bank_rest.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserRequestDto {
    private String username;
    private String role;
    @NotBlank
    private String password;
}